requirements.txt
pinecone
openai
langchain
python-dotenv
sentence-transformers
fastapi
streamlit